requirements.txt
python-decouple
# telethon
pyrogram

paypalrestsdk
aiohttp 